Divide by explicitly declared float

+++ src/porting_android.cpp
@@ -386,7 +386,7 @@ void upgrade(const std::string &item)
 float getTotalSystemMemory() {
 	long pages = sysconf(_SC_PHYS_PAGES);
 	long page_size = sysconf(_SC_PAGE_SIZE);
-	float divisor = 1024 * 1024 * 1024;
-	return pages * page_size / divisor;
+	int divisor = 1024 * 1024 * 1024;
+	return pages * page_size / float(divisor);
 }
 }
